feat: Add GooglePay payment method model
The commit adds the `GooglePay` payment method model to the `PayPalKit` project. This model represents a payment method that allows buyers to pay using Google Pay. It includes properties such as the name and email address of the account holder, phone number, payment card information, decrypted token details, and additional attributes associated with the use of this card.