feat: Add AuthorizeOrderRequest and AuthorizeOrderResponse models
This commit adds the `AuthorizeOrderRequest` and `AuthorizeOrderResponse` models to the `PayPalKit` project. The `AuthorizeOrderRequest` model represents the request body for authorizing payment for an order, while the `AuthorizeOrderResponse` model represents the response received after authorizing payment for an order. These models enhance the functionality of the PayPalKit by providing structured data for the authorization process.