feat: Add CaptureOrderRequest model and API route
This commit adds the `CaptureOrderRequest` model and the corresponding API route in the `PayPalOrders.swift` file. The `CaptureOrderRequest` model represents the request body for capturing payment for an order. The API route allows the capturing of payment for an order by sending the `CaptureOrderRequest` object. This addition enhances the functionality of the PayPalKit.