Implement Flexible Authentication for PayPal API Client (#4)
* feat: Rename PayPalOrders to Routes/PayPalOrders.swift and remove commented code

* feat: Add BinDetails model for Bank Identification Number (BIN) details

* feat: Add properties for card details in Bancontact model

* chore: update OrderDetails model

* refactor: reorganize folder structure for better maintainability

- Moved core API handling files to a new Core directory
- Grouped all data models under a single Models directory
- Organized PaymentMethods by categories (DigitalWallets, BankTransfers, Cards)

* feat: Update createOrder method return type to OrderDetails

* feat: Add BIC property to BankTransfer payment method models

* feat: Add Swift Crypto dependency for PayPalKit

* feat: Refactor error handling in Request

The commit refactors the error handling in the `Request.swift` file of the `PayPalKit` project. It replaces the custom `SubscriptionError` enum with a new `RequestError` enum to handle JSON encoding, decoding, and body extraction failures. This change improves the clarity and maintainability of the code.

* feat: Update createOrder method return type to OrderDetails

* refactor: folder structure

* feat: Add ConfirmPaymentSourceRequest model and API route

This commit adds the `ConfirmPaymentSourceRequest` model and the corresponding API route in the `PayPalOrders.swift` file. The `ConfirmPaymentSourceRequest` model represents the request body for confirming the payment source for an order. The API route allows the confirmation of the payment source for an order by sending the `ConfirmPaymentSourceRequest` object. This addition enhances the functionality of the PayPalKit.

* feat: Add AuthorizeOrderRequest and AuthorizeOrderResponse models

This commit adds the `AuthorizeOrderRequest` and `AuthorizeOrderResponse` models to the `PayPalKit` project. The `AuthorizeOrderRequest` model represents the request body for authorizing payment for an order, while the `AuthorizeOrderResponse` model represents the response received after authorizing payment for an order. These models enhance the functionality of the PayPalKit by providing structured data for the authorization process.

* feat: Add CaptureOrderRequest model and API route

This commit adds the `CaptureOrderRequest` model and the corresponding API route in the `PayPalOrders.swift` file. The `CaptureOrderRequest` model represents the request body for capturing payment for an order. The API route allows the capturing of payment for an order by sending the `CaptureOrderRequest` object. This addition enhances the functionality of the PayPalKit.

* feat: Add PatchOperation model for updating orders

This commit adds the `PatchOperation` model to the `PayPalKit` project. The `PatchOperation` model represents a patch operation for updating orders. It includes properties such as `op` (operation type), `path` (target location), `value` (new value), and `from` (source location for move and copy operations). This addition enhances the functionality of the PayPalKit by providing a structured way to update orders.

* refactor: Update PayPalAPIHandler to use PayPalAuthType for authentication

This commit refactors the `PayPalAPIHandler` struct in the `PayPalKit` project to use the `PayPalAuthType` enum for authentication instead of a single access token. The `PayPalAuthType` enum provides options for authentication using either a client secret or a bearer token. This change enhances the flexibility and security of the authentication mechanism in the PayPalKit.

* refactor: Update PayPalClient initialization to use PayPalAuthType for authentication